refactor header file
